--- integration_tests/.gitignore
@@ -1,21 +1,0 @@
-# Distribution / packaging
-.Python
-*.pyc
-env/
-build/
-develop-eggs/
-dist/
-downloads/
-eggs/
-.eggs/
-lib/
-lib64/
-parts/
-sdist/
-var/
-*.egg-info/
-.installed.cfg
-*.egg
-
-# Serverless directories
-.serverless
