--- integration_tests/serverless.yml
@@ -1,74 +1,0 @@
-service: aws-lambda-cache
-
-custom:
-  parameter_name: /lambda_cache/something
-  parameter_name_2 : /lambda_cache/test/something_else
-  secure_string: /lambda_cache/test/secure/somethingsecure
-  string_list: /lambda/cache/test/somelist
-
-provider:
-  name: aws
-  runtime: python3.6
-  stage: test
-  region: ap-southeast-1
-  iamRoleStatements:
-    - Effect: Allow
-      Action:
-        - ssm:GetParameter
-      Resource:
-        - Fn::Join:
-          - ":"
-          - - "arn:aws:ssm"
-            - Ref: AWS::Region
-            - Ref: AWS::AccountId
-            - "parameter${self:custom.parameter_name}"
-        - Fn::Join:
-          - ":"
-          - - "arn:aws:ssm"
-            - Ref: AWS::Region
-            - Ref: AWS::AccountId
-            - "parameter${self:custom.parameter_name_2}"
-        - Fn::Join:
-          - ":"
-          - - "arn:aws:ssm"
-            - Ref: AWS::Region
-            - Ref: AWS::AccountId
-            - "parameter${self:custom.secure_string}"
-        - Fn::Join:
-          - ":"
-          - - "arn:aws:ssm"
-            - Ref: AWS::Region
-            - Ref: AWS::AccountId
-            - "parameter${self:custom.string_list}"
-    - Effect: Allow
-      Action:
-        - kms:Decrypt
-      Resource:
-        Fn::Join:
-          - ":"
-          - - "arn:aws:kms"
-            - Ref: AWS::Region
-            - Ref: AWS::AccountId
-            - "alias/aws/ssm"  # Default AWS Managed CMK Key
-
-package:
- include:
-   - ../aws_lambda_cache/**
-   - ../tests/variables_data.py
- exclude:
-   - ../aws_lambda_cache/__pycache__/**
-
-functions:
-  single_handler:
-    handler: test_ssm.single_parameter
-  multi_handler:
-    handler: test_ssm.multi_parameter
-  default_param:
-    handler: test_ssm.rename_param
-  secure_string:
-    handler: test_ssm.secure_string
-  string_list:
-    handler: test_ssm.string_list
-  assign_parameter:
-    handler: test_ssm.assign_parameter
-
