+++ simple_lambda_cache/secret_manager.py
@@ -1,7 +1,7 @@
 import functools
 import boto3
 
-from .caching_logic import check_cache, get_entry_name
+from .caching_logic import get_decorator, get_value, get_entry_name
 from .exceptions import ArgumentTypeNotSupportedError
 
 
@@ -19,23 +19,10 @@ def secret_cache(name, ttl_seconds=60, entry_name=False):
 
     """
 
-    def decorator(func):
-        @functools.wraps(func)
-        def inner_function(event, context):
-
-            response = check_cache(
-                argument=name,
-                ttl_seconds=ttl_seconds,
-                entry_name=entry_name,
-                miss_function=get_secret_from_secrets_manager,
-            )
-            # Inject {parameter_name: parameter_value} into context object
-            context.update(response)
-
-            return func(event, context)
-
-        return inner_function
-
+    decorator = get_decorator(argument=name,
+                              ttl_seconds=ttl_seconds,
+                              entry_name=entry_name,
+                              miss_function=get_secret_from_secrets_manager)
     return decorator
 
 
@@ -44,22 +31,19 @@ def get_secret_cache(name, ttl_seconds=60, entry_name=False):
     Wrapper function for parameter_caching
 
     Args:
-        name(string): Name of the parameter in System Manager Parameter Store
+        name(string): Name of the Secret in Secrets Manager (arn is also accepted)
         ttl_seconds(int) : Time to Live of the parameter in seconds
         var_name(string) : Optional name of parameter to inject into context object
 
     Returns:
-        parameter_value(string)  : Value of the parameter
+        secret_value(string)  : Value of the parameter
     """
 
-    response = check_cache(
-        argument=name,
-        ttl_seconds=ttl_seconds,
-        entry_name=entry_name,
-        miss_function=get_secret_from_secrets_manager,
-    )
-    parameter_value = list(response.values())[0]
-    return parameter_value
+    secret_value = get_value(argument=name,
+                                ttl_seconds=ttl_seconds,
+                                entry_name=entry_name,
+                                miss_function=get_secret_from_secrets_manager)
+    return secret_value
 
 
 def get_secret_from_secrets_manager(name):
